various
Essentially was trying to fix a bug whereby on collision with enemy the game crashes. I suspect it's to do with ordering but I don't know how that works in Godot yet as the tutorial didn't cover it.

Not really crucial to fix as I was trying to introduce a feature where you can press X to disable sounds, which does work for jumping sound but crashes on enemy related death. It's to do with how the killzone was created for 100% sure but again, as it was a tutorial and details were not covered, it's not something I can figure out yet, nor is it worth any further effort.